Support for updating the filters of the Filter Tab.
When filters are created and the logs are extracted based on the filter to the new Tab,
the filters can not be updated for now so the logs are fixed by the filters. However
sometimes there might be need to change some filter fields, or add some new filter
patterns, we can just update the existing Tab instead of creating a new filter Tab.

Signed-off-by: CoryXie <cory.xie@gmail.com>